🐛 fix: code-review findings on batch 4 (#1448)
Two reviewers, split by scope. Shard 1 PASS (2 warnings), shard 2 FAIL
(6 warnings). All applied. Two were arithmetic, and both were errors in
prose this branch wrote while retiring an earlier false claim.

The side condition said inversion needs a ground darker than BOTH tokens.
It does not: it begins in the band BETWEEN them. The arm's own ground
(#666666) sits there — darker than muted, lighter than inkSecondary — so
the prose described neither the mechanism nor its own counterexample. That
is the more dangerous shape of the two, because "my ground is not as dark
as the replacement, so I'm safe" reads as diligence and lands wrong.

And "raises the ratio only while the ground is lighter than muted" was
written as necessary when it is sufficient — a ground just below muted's
luminance still raises it. The operative rule is unchanged and stays
conservative; a branch retiring one false unconditional should not install
another.

Also applied:
- The ADR preamble and roster row still carried the pre-risk-review
  framing (direction instead of measurement, "pinned by symbol rather than
  a fixture arm") — true of batches 1 and 3, not of this one.
- The amendment cited ledger §3.3's unconditional clause in the present
  tense, two paragraphs above refuting it. Past-tensed.
- SimulationScrimStyle.fill was called a retained row. It is a fill; the
  row is SimulationView · loading-scrim subtitle — and since the scrim is
  a near-black, that row is the one shipped ground plausibly inside the
  inversion band. §5 now carries the warning rather than leaving it in a
  test's doc comment.
- "kept" used where the branch reserves "retained".
- MutedSweepLedgerTests' own prose still said post-batch-3; batch 4 is the
  first to decrement an entry rather than empty it.
- The BatchFour window doc justified its width with a false uniqueness
  claim (return Color.mossInk IS unique, two lines up). Replaced with the
  real reason — anchoring on a sibling arm couples the pin to .completed's
  routing — and the two residuals of the wide window are now stated.
- The ModelRow after-figure asserts its before-figure rows exist rather
  than citing them, since the composite params are re-typed.

The last of those pushed +MutedTranscript past file_length. Squeezing
prose to sit exactly on 400 is the wrong answer, so the two batch-4 arms
move to +MutedDirection: neither transcribes a doc figure, and neither is
reached by check-measurement-transcripts.py's anchors.

3554 tests green; swiftlint --strict, --check, --census and the mossink
membership gate all clean.

Part of #1448
