+++ Pastura/PasturaTests/Views/MutedSweepLedgerTests.swift
@@ -13,7 +13,11 @@ import Testing
 /// they take opposite remedies**: a rename or a deletion means re-anchor the
 /// row by symbol rather than dropping it; an **applied batch** means the row
 /// stays put with its `B` marker bolded and the entry here is dropped, which is
-/// what batches 1, 5, 2 and 3 did. Read the file's § 5 rows before choosing.
+/// what batches 1, 5, 2, 3 and 4 did. Read the file's § 5 rows before choosing.
+/// A batch can also **decrement** an entry rather than drop it — batch 4 took
+/// `GameHeaderStatus.swift` 2 → 1, moving the label while the wash kept the
+/// spelling — so a count that fell without reaching zero is the same question,
+/// not a different one.
 ///
 /// **It fires on an addition too**, which is the half prose cannot do. The
 /// expectation is a whole-map comparison, so a new file reaching for
@@ -42,8 +46,10 @@ struct MutedSweepLedgerTests {
   /// the spelling on the glyph above — and emptied none. Batch 3 repointed six
   /// sites in three files — the eliminated rows in `SimulationResultCard` /
   /// `ScoreboardSheet` and the prediction countdown — and emptied none either.
-  /// So this table is the post-batch-3 population, not § 1's `9a40565a`
-  /// baseline.
+  /// Batch 4 repointed three — the header pill's terminal-state labels, the
+  /// model row's meta, and the blocked clear-all label — emptying two files and
+  /// decrementing a third. So this table is the post-batch-4 population, not
+  /// § 1's `9a40565a` baseline.
   private static let expectedMutedOccurrences: [String: Int] = [
     "Views/Community/SharedScenarios/GalleryCatalogRow.swift": 1,
     "Views/Community/SharedScenarios/GalleryScenarioDetailView+Highlight.swift": 2,
