+++ docs/design/design-system.md
@@ -864,7 +864,8 @@ Sim 画面に限った2つの例外的コンポーネント。Source: `Simulatio
 - **置き換えるトークンは「地を所有するファミリ」が供給する。** §8 が縛るのは**比**であって特定のトークンではない。不透明な `--moss-soft` 上なら §2.3/§2.6 の Soft+Ink ペアリングで `--moss-ink`、中立なカード面なら `--ink-2`（`PasturaApp` の DB 移行失敗文が先例）、**半透明のウォッシュ上なら §2.2/§2.3 の `*-on-wash` 役割トークン** — 苔系は `--moss-on-wash`、インク系は `--ink-on-wash`。この 3 つ目の地を落とすと、半透明のインク系ウォッシュを追ってきた読者が `--ink-2` に着地する — つまり dark で 4.413〜4.773 を出した当のトークンで、#1408 はまさにそれを直した（導出は ADR-028 § Amendment 2026-08-13（#1408））。**§2.4 の L3 プリセットの段を借りないこと** — 値が合っても §2.4 は DL 進捗の役割を持つ梯子で、借りると2つのファミリが結合する（ADR-028 § "Three narrower rejections"）。同じ #4A4E3D が別の役割で要るとき、このリポジトリは `--header-meta-ink` を**別トークンとして起こした**（§2.12）。L3 が正解なのは §2.4 自身のメタ面。導出は ADR-028 § Amendment 2026-08-13（#1427）
 - ⚠️ **静的な比が出ない地でも routing は決まる。ただし「測れない」を最初の答えにしないこと。** マテリアル（`.ultraThinMaterial` / `.regularMaterial`）、半透明の上にさらに半透明が重なった層、Pastura トークンを持たないシステムのシート地 — この 3 種は真の地が実行時に決まる。**それでも**地を所有するファミリは読めるので、置き換え先の既定は上のバレットのまま: 中立なら `--ink-2`、インク系の半透明ウォッシュなら `--ink-on-wash`、苔系なら `--moss-on-wash`。根拠の付け方には**優先順位がある**:
   1. **nominal 地で実測する（既定）。** 不透明な下限が 1 つでも書けるなら、それを地として fixture で測り、その行に *nominal*（worst case ではない）と注記する。`GameHeader` のバーは `--screen-bg@0.78` をマテリアルの上に敷くので下限が書け、`DesignTokensTests+MossOnWash` の `GameHeaderStatus.active` 行が実際にこの形で測られている。**同じピルの終端状態アームも同じ形で測る**（#1448 batch 4）。nominal な比は bound ではないので、埋めるのは実機 QA。
-  2. **それも書けないときに限り、方向で論証する。** 置き換え先が light で `--muted` より暗く、dark で `nightMuted` より明るいこと。⚠️ **これだけでは足りない** — 「暗くすれば比が上がる」は**地が `--muted` より明るいときにしか成り立たない**。地が両方の前景より暗いと逆転し、repoint はコントラストを**下げる**（`--muted` のほうが地から遠くなるため）。反例は `DesignTokensTests+MutedTranscript` の `directionArgumentInvertsOnAGroundDarkerThanMuted` が構成して pin しているので、比はここに写さない。だから方向論証は「地の明度が light では `--muted` より上、dark では `nightMuted` より下だと**言い切れる**場合」に限って使うこと。言い切れないなら 1 に戻るか、repoint 自体を見送る。使ったサイトは **why-comment に「方向論証」と明記し、比は書かない**（無い）。ガードは fixture ではなく `MutedSweepLedgerTests` の**シンボル固定**。先例: #1448 batch 1 の `DLCompleteOverlay`（マテリアルのみで下限が無い）、batch 3 の `ScoreboardSheet` の 2 行（システムのシート地）。
+  2. **それも書けないときに限り、方向で論証する。** 置き換え先が light で `--muted` より暗く、dark で `nightMuted` より明るいこと。⚠️ **これだけでは足りない** — 「暗くすれば比が上がる」が**保証されるのは、地が `--muted` より明るいときだけ**。地がそこより暗くなると、**2つの前景の中間のどこかで符号が反転し**、そこから下では repoint がコントラストを**下げる**（`--muted` のほうが地から遠くなるため）。⚠️ **反転が起きるのは「地が両方の前景より暗いとき」ではなく、その手前の帯である** — 「置き換え先より暗くはないから安全」という読み方は成立しない。反例は `DesignTokensTests+MutedDirection` の `directionArgumentInvertsOnAGroundDarkerThanMuted` が構成して pin しており、その地はまさにこの帯（`--muted` より暗く `--ink-2` より明るい）にある。比はここに写さない。だから方向論証は「地の明度が light では `--muted` より上、dark では `nightMuted` より下だと**言い切れる**場合」に限って使うこと — 帯に入りうるなら使えない。言い切れないなら 1 に戻るか、repoint 自体を見送る。使ったサイトは **why-comment に「方向論証」と明記し、比は書かない**（無い）。ガードは fixture ではなく `MutedSweepLedgerTests` の**シンボル固定**。先例: #1448 batch 1 の `DLCompleteOverlay`（マテリアルのみで下限が無い）、batch 3 の `ScoreboardSheet` の 2 行（システムのシート地）。
+
   ⚠️ **どちらの経路も「測れないから上げてよい」ではない** — repoint の要否は先に上の役割判定（アンビエントか、A1〜A5 か）で決めること。ここが効くのは「上げると決めた後、どのトークンへ、どんな根拠で」だけで、役割がアンビエントなら測れない地でも据え置く
 - **§2.7 の**値**を使った塗りでも、状態で出入りするものは §2.3 のウォッシュではない。** `hover` / `pressed` / `selected`（moss 6/12/18%）を**要素の状態に応じて出したり消したりする**塗りは、§2.7 自身が「下地のサーフェス色を問わない」と書くとおり状態を示す薄膜であって、要素の地を定義しない。したがって既定の「半透明のウォッシュ上なら `*-on-wash`」は**そこには及ばない** — 地はオーバーレイの下のサーフェスのままで、文字トークンはその地で選ぶ。**独立した根拠**: ウォッシュと読むと同じ 1 行の同じ文字が選択時 `--moss-on-wash` / 非選択時 `--ink-2` と**実行時に入れ替わる**ことになり、routing として成立しない。実例: `ModelRow` の選択行は `moss@0.06`（=`hover` と同値）の下が `--bubble-bg` で、タイトルは `--ink`、メタは `--ink-2`（#1448 batch 4）。⚠️ **これはトークン単位の判定ではない。** 同じ §2.7 トークンを**常設**の装飾ウォッシュとして塗るサイトは既定どおりウォッシュで、`--moss-on-wash` が要る — `GalleryCatalogRow.categoryChip` が `Color.selected` をそう使っており、#1327 が実測して `mossWashSites` に pin している。**見分けるのは塗りの出入りであってトークン名ではない**。ただし**比はオーバーレイ込みで測る**こと（薄膜の分だけ下がるので、素の地の値を流用しない）。fixture の `composite` を 2 段重ねれば出る
 - **例外は1つだけ。しかも「役割が文書化されていれば既定を上書きしてよい」ではない。** 上の既定（半透明ウォッシュ → `*-on-wash`）を外れてファミリの **Ink 段**を使ってよいのは、次の4つを**すべて**満たす場合に限る: **(1)** §2.2/§2.3 がその Ink 段に**その意味そのもの**の役割を割り当てている、**(2)** 同じファミリのウォッシュ上でその段を描いている出荷済みのピルが既にある — **ただしこの例外を経由して着地したサイトは先例に数えない**（さもないと最初の1件が以後すべての (2) を自動的に満たし、条件がラチェットで緩む）、**(3)** その要素が**行の主役を上回らない**（階層が反転しない）、**(4)** 借りる Ink 段が、過去の amendment でコントラスト不足を理由に退けられたトークン**ではない**。今日これを満たすのは**完了**だけで、答えは `--moss-ink` — (1) は §2.3 の「完了タイトル」、(2) は `ResultsView` の完了ピル、(3) はヘッダのタイトルが完了ピルを上回ること（light 13.147 対 8.604、dark 10.769 対 6.047）、(4) は `--moss-ink` がどの amendment でも失敗トークンでないこと。`GameHeaderStatus.completed` がこの経路
